LayoutTests/fast/dom/Element/contains-method-expected.txt

This test checks behavior of Element.contains.

[test1 [test 2]]
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.contains is undefined.
PASS test1.contains(test2) is true
PASS test1.contains(test1TextChild) is false
PASS test1.contains(123) is false
PASS test1.contains(null) is false
PASS successfullyParsed is true

TEST COMPLETE
